requirements.txt

numpy
pandas
tqdm
pyyaml
regex

opencv-python
pillow
scikit-image

yt-dlp
webvtt-py


paddlepaddle
paddleocr
pytesseract

ultralytics

scenedetect

nltk

ollama
streamlit
